backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic[email]
sqlite-utils
python-dotenv
alembic
passlib
httpx
pytest
python-multipart
bcrypt
python-jose[cryptography]==3.3.0
ldap3
bcrypt
